Ch. 17: Change favorite checkbox to use AsyncTask.
All database code should be run the background. As an example, the
code for updating the favorite column in the Drink table in the database
by clicking the checkbox in DrinkActivity has been changed to an
AsyncTask that runs in the background. The app still functions as
before, but now that database update runs on a background thread instead
of the main thread. Processing of views must still be done on the main
thread, as interacting with them in a background thread will cause
exceptions.